+++ php/tests/Application/Command/VariableTypesTest/InstanceofComplexIfBooleanAnd.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace A;
+
+if ((1 ^ 0) && true && $b instanceof B && ($test || false && true)) {    
+    // <MARKER>
+}