+++ php/tests/Application/Command/VariableTypesTest/InstanceofComplexIfLogicalAnd.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace A;
+
+if ((1 ^ 0) and true and $b instanceof B and ($test || false and true)) {    
+    // <MARKER>
+}